Resolve `this.handlersDirectory` relative to `process.cwd()`.
The original implementation was causing issues in my `yarn` workspaces, hoisted modules, environment as it supposed that my service had a local installation of `serverless-offline-step-functions`, which it didn't.

I mean, ideally we name space `this.handlersDirectory` on the name of the serverless service so multiple services can support sharing the same module instance at the same time, but that's beyond my current use case. 😅